Rename scripts to include xgboost
